software/user/libm2sdr: split RFIC layer into backend interface
