+++ litex_m2sdr/software/user/Makefile
@@ -44,6 +44,8 @@ LIBM2SDR_OBJS = \
 	libm2sdr/m2sdr_ad9361_spi.o \
 	libm2sdr/m2sdr_flash.o \
 	libm2sdr/m2sdr_device.o \
+	libm2sdr/m2sdr_rfic_registry.o \
+	libm2sdr/m2sdr_rfic_ad9361.o \
 	libm2sdr/m2sdr_stream.o \
 	libm2sdr/m2sdr_rf.o \
 	libm2sdr/m2sdr_utils.o \
